Auto create GuildConfig using signals to avoid circlular imports
